move shopClass to extern file
