+++ lib/page-encointer/bazaar/shop/shopClass.dart
@@ -1,0 +1,15 @@
+class Shop {
+  final String name;
+  final String description;
+  final String imageHash;
+
+  Shop({this.name, this.description, this.imageHash});
+
+  factory Shop.fromJson(Map<String, dynamic> json) {
+    return Shop(
+      name: json['name'],
+      description: json['description'],
+      imageHash: json['imageHash'],
+    );
+  }
+}
